+++ contracts/autostake-features/LimitedAutoStake.sol
@@ -10,7 +10,7 @@ contract LimitedAutoStake is AutoStake {
 
 	uint256 public immutable stakeLimit;
 
-	constructor(address token, uint256 _throttleRoundBlocks, uint256 _throttleRoundCap, uint256 stakeEnd, uint _stakeLimit) public AutoStake(token, _throttleRoundBlocks, _throttleRoundCap, stakeEnd) {
+	constructor(address token, uint256 _throttleRoundBlocks, uint256 _throttleRoundCap, uint256 stakeEnd, uint256 _stakeLimit) public AutoStake(token, _throttleRoundBlocks, _throttleRoundCap, stakeEnd) {
 		require(_stakeLimit != 0 , "LimitedAutoStake:constructor::stake limit should not be 0");
 		stakeLimit = _stakeLimit;
 	}
